requirements.txt
asgiref==3.3.2
django==3.1.7
django-filter==2.4.0
django-rest-passwordreset==1.1.0
psycopg2-binary==2.8.6
djangorestframework==3.12.4
django-cors-headers==3.6.0
django-simple-history==2.12.0
django-extensions==3.1.0
django-cacheops==5.1
pyjwt==1.7.1
dateutils==0.6.12
pytz==2021.1
sqlparse==0.4.1
django-rest-swagger==2.2.0
pillow==8.4
